Use BACNET_ENGINEERING_UNITS for basic object API for units properties (#1104)

+++ src/bacnet/basic/object/piv.c
@@ -239,8 +239,8 @@ int PositiveInteger_Value_Read_Property(BACNET_READ_PROPERTY_DATA *rpdata)
             break;
 
         case PROP_UNITS:
-            apdu_len =
-                encode_application_enumerated(&apdu[0], CurrentAV->Units);
+            apdu_len = encode_application_enumerated(
+                &apdu[0], (uint32_t)CurrentAV->Units);
             break;
             /* BACnet Testing Observed Incident oi00109
                     Positive Integer Value / Units returned wrong datatype -
